add unit tests

+++ txnbuild/manage_buy_offer.go
@@ -69,7 +69,7 @@ func (mo *ManageBuyOffer) FromXDR(xdrOp xdr.Operation) error {
 	mo.OfferID = int64(result.OfferId)
 	mo.Amount = amount.String(result.BuyAmount)
 	if result.Price != (xdr.Price{}) {
-		mo.Price = price.StringFromFloat64(float64(result.Price.N / result.Price.D))
+		mo.Price = price.StringFromFloat64(float64(result.Price.N) / float64(result.Price.D))
 	}
 	buyingAsset, err := assetFromXDR(result.Buying)
 	if err != nil {
